refactor: split four oversized worker and Workshop modules
meshImprint, baseplateConnectors, splitBinBuilder and WorkshopScene each carried
more than 500 lines of code. Each now delegates a self-contained slice to a
sibling module: the tool preparation and frame math in meshImprintTools, the
connector solid builders in connectorShapes, the piece export and preview
tessellation in splitPieceExport, and the R3F bridge and catch-plane components
in WorkshopSceneBridges and WorkshopCatchPlanes. Public names re-export from
their original modules, so callers and tests are unchanged.
